[#153] Upgraded dependencies
Problem: there have been recent updates to the TZIP-16 spec and the
morley-metadata package, so our implementation is now outdated.

Solution: update to the latest version of morley-metadata. This implies
also updating morley/lorentz/indigo.

* We no longer need some of the tests we had in the
  `Lorentz.Contracts.Test.TZIP16` module, because `morley-metadata` now
  has its own tests that cover this.
* Because morley-metadata now uses TH to check a view's code, we now
  have to have the contract's type definitions and the contract's
  metadata+views in separate modules (due to the GHC stage restriction).
  I've moved the type definitions to their own `.Types` module, and in
  the next commit I'll move the metadata+views to their own `.Metadata`
  module. I'm doing this in 2-steps to make it easier to review.
