+++ haskell/nettest/FA1_2Comparison.hs
@@ -6,6 +6,7 @@ module FA1_2Comparison
   ) where
 
 import qualified Data.Map as Map
+import Fmt (build)
 
 import Lorentz (BigMap(..), EntrypointRef(Call), TAddress(..), toVal)
 import Michelson.Typed (convertContract, untypeValue)
@@ -49,7 +50,9 @@ fa1_2ComparisonScenario = uncapsNettest $ do
         , rPendingOwner = Nothing
         }
 
-  cmrFA1_2Address <- nettestOriginateContractMetadataContract (SFA2.metadataJSON Nothing)
+  metadata <- either (failure . build) pure $ SFA2.metadataJSON Nothing
+
+  cmrFA1_2Address <- nettestOriginateContractMetadataContract metadata
   let fa1_2Storage = SFA1_2.Storage
         { sDefaultExpiry = 1000
         , sLedger = BigMap balances
@@ -70,7 +73,7 @@ fa1_2ComparisonScenario = uncapsNettest $ do
   comment "Calling transfer"
   callFrom (AddressResolved owner1) fa1_2ContractAddr (Call @"Transfer") (#from .! owner1, #to .! owner2, #value .! 2)
 
-  cmrAddress <- nettestOriginateContractMetadataContract (SFA2.metadataJSON Nothing)
+  cmrAddress <- nettestOriginateContractMetadataContract metadata
   let fa2Storage = SFA2.Storage
         { sDefaultExpiry = 1000
         , sLedger = BigMap balances
