+++ haskell/src/Stablecoin/Client/Main.hs
@@ -6,6 +6,7 @@ module Stablecoin.Client.Main
   , mainProgramIO
   ) where
 
+import Control.Exception.Uncaught (displayUncaughtException)
 import Data.Coerce (coerce)
 import Data.Map (lookup)
 import Fmt (pretty)
@@ -15,7 +16,6 @@ import Morley.Client
   TezosClientError(UnknownAddressAlias), mkMorleyClientEnv, rememberContract, runMorleyClientM)
 import Morley.Client.TezosClient (resolveAddress)
 import qualified Options.Applicative as Opt
-import Util.Exception (displayUncaughtException)
 import Util.Named ((.!))
 
 import Stablecoin.Client.Contract (InitialStorageData(..))
