+++ nix/sources.json
@@ -5,10 +5,10 @@
         "homepage": "",
         "owner": "input-output-hk",
         "repo": "hackage.nix",
-        "rev": "9d285f2e91105e47222ea49baa5892ec3134fc65",
-        "sha256": "0byx9v4fcd89f4x591qdicxd1hyd5jnrad7sdvja5pk06knz57h8",
+        "rev": "8ab47a84664f99da20ee5b956851a2431b9073d7",
+        "sha256": "1i4vh3svcj7r68ykw0six7wvdf8q6k4nfb07gvaqhh53qbm97vlc",
         "type": "tarball",
-        "url": "https://github.com/input-output-hk/hackage.nix/archive/9d285f2e91105e47222ea49baa5892ec3134fc65.tar.gz",
+        "url": "https://github.com/input-output-hk/hackage.nix/archive/8ab47a84664f99da20ee5b956851a2431b9073d7.tar.gz",
         "url_template": "https://github.com/<owner>/<repo>/archive/<rev>.tar.gz"
     },
     "haskell-nix-weeder": {
